# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tools API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Rota para realizar o login e obter a token access",
                "operationId": "c388b5cc7b4f41f4892c8c9a5777ec25",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "sucesso, retorno do token de acesso"
                    },
                    "401": {
                        "description": "Erro ao efetuar o login"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Sair do login",
                "operationId": "01860934654856fdd10b21c4bbacceb7",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tools": {
            "get": {
                "tags": [
                    "/tools"
                ],
                "description": "Lista de ferramentas",
                "operationId": "e270363ef10f02a66a7b29d3a9ca6358",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "filtrar ferramentas utilizando uma busca por tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de ferramentas"
                    },
                    "404": {
                        "description": "Nenhuma ferramenta foi encontrado no banco de dados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/tools"
                ],
                "description": "Criação de nova ferramenta",
                "operationId": "3efaf0e49bca0f4489cc82a085175e71",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "link": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opecação bem sucedida. Ferramenta criada com sucesso"
                    },
                    "400": {
                        "description": "Sintaxe JSON inválida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tools/{id}": {
            "get": {
                "tags": [
                    "/tools"
                ],
                "description": "Exibir detalhes de uma ferramenta",
                "operationId": "0207264489549630a33152d9563d3343",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Realizar busca pelo id da ferramenta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da ferramenta"
                    },
                    "404": {
                        "description": "Nenhuma ferramenta foi encontrado no banco de dados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/tools"
                ],
                "description": "Excluir ferramenta",
                "operationId": "4706b2f2cc973eae14cfb39981466454",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Realizar a exclusão pelo id da ferramenta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        },
        "0": []
    }
}